Complete CLR integration - fix compile errors and graph relationships
ConceptDiscoveryRepository:
- Add System.Data, Microsoft.Data.SqlClient using directives
- Define ClrConceptResult DTO matching fn_DiscoverConcepts schema
- Add binary conversion helpers (ConvertVarbinaryToDoubleArray, ConvertDoubleArrayToVarbinary)
- Replace SqlQueryRaw with direct SqlCommand (better control over VARBINARY handling)
- Fix BindConceptsAsync to use VECTOR_DISTANCE via raw SQL (remove EF Core materialization)
- Fix CreateConceptRelationshipsAsync to use graph.AtomGraphNodes/Edges with proper node creation
- Remove IS_MEMBER_OF edge table reference (doesn't exist in schema)

AutonomousAnalysisRepository:
- Cast long to int for DurationMs (PerformanceAnomaly.DurationMs is int)

AutonomousActionRepository:
- Remove unused ConvertSqlVectorToDoubleArray helper (no longer needed)

All repositories now delegate to SQL Server:
- CLR functions for compute (DBSCAN, IsolationForest, VECTOR_DISTANCE)
- Stored procedures for multi-step workflows (sp_HybridSearch, sp_SpatialVectorSearch)
- Raw SQL for graph operations (MATCH syntax, node/edge insertion)

Architecture principle enforced: Database is the AI runtime, C# is thin orchestration.

+++ src/Hartonomous.Data/Repositories/AutonomousAnalysisRepository.cs
@@ -152,7 +152,7 @@ public class AutonomousAnalysisRepository : IAutonomousAnalysisRepository
                 {
                     InferenceRequestId = inferenceId,
                     ModelId = modelId,
-                    DurationMs = durationMs,
+                    DurationMs = (int)durationMs,
                     AvgDurationMs = avgDurationMs,
                     SlowdownFactor = durationMs / avgDurationMs
                 });
